requirements.txt
python-telegram-bot
uuid
json
datatime
time
requests